feat(payments): refactor payment dashboard into modular components and improve modal management
- broke page.jsx into multiple parts
- handled proper props passing to each component
